requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
Flask==3.0.3
fpdf==1.7.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
mysql-connector-python==9.0.0
passlib==1.7.4
pillow==11.0.0
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python-multipart==0.0.10
pytweening==1.2.0
pywhatkit==5.4
pywin32==308
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
wikipedia==1.4.0
